fixed problem with document descriptor inialization
ATTENTION: This commit is interesting because it fixes
a real nasty bug. The InitializeDocumentDescriptorAsAttachCommandHandler
had a bug.

It uses FindAndModify to inizialize a document descriptor then
it call some other Repository.Save functions to create links.

FindAndModify did not check for already initialized document descriptor
and this could potentially create this subtle bug under load

1) command handler started, documentdescriptor is initialized
2) HAndle is linked, but NEventStore mongo persister throw a
   concurrent exception
3) Base handler retry to execute the command
4) command handler try to initialize again document descriptor
   a) This raises a Domain Exception
   b) This stops further command execution
   c) Document (handle) is not linked to this descriptor.