Better handling of queue when docs are deleted
Closes #15.

When a document descriptor is deleted the queue
manager now deletes all the jobs related to that
descriptor.

This is done to avoid that jobs run against a
deleted document descriptor, doing unnecessary
work. Also added a check in add format to document
to verify if the documentdescriptor is really deleted
to gave a better error message.