Changed logic in ComposePdf to wait for pdf format
Now PdfComposer job, when it does not find a pdf
version for an handle, ask for all the job queued
for that specific handle.

If PdfConverter job is not run, or the file has
an office extension and the office queue still
did not run, the job will wait.

Previous logic was flawed, because if the queue
manager did not generated job when PdfComposer check
for job, it incorrectly believe that the file has not
pdf format.