Fixed get info for handle.
If you have a DocumentDescriptor that points
to more than one handle (de-duplication) if
you search for handle, only the first one
is returned. This is not correct because
when we are searching for document handle
we do not care about de-duplication, but
we wish to know all handles that matches
the query.